requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Gunicorn==21.2.0
psycopg2-binary==2.9.6
google-generativeai==0.3.2
google-api-python-client==2.108.0
google-auth==2.23.4
